+++ packages/wonder-blocks-form/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @khanacademy/wonder-blocks-form
 
+## 7.1.5
+
+### Patch Changes
+
+- Updated dependencies [3dc5dac]
+    - @khanacademy/wonder-blocks-tokens@5.2.0
+    - @khanacademy/wonder-blocks-clickable@6.1.5
+    - @khanacademy/wonder-blocks-layout@3.1.5
+
 ## 7.1.4
 
 ### Patch Changes