+++ packages/wonder-blocks-toolbar/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @khanacademy/wonder-blocks-toolbar
 
+## 5.1.5
+
+### Patch Changes
+
+- Updated dependencies [3dc5dac]
+    - @khanacademy/wonder-blocks-tokens@5.2.0
+
 ## 5.1.4
 
 ### Patch Changes